+++ .phan/baseline.php
@@ -11,17 +11,13 @@ return [
 	// # Issue statistics:
 	// PhanPluginDuplicateAdjacentStatement : 10+ occurrences
 	// MediaWikiNoEmptyIfDefined : 8 occurrences
-	// SecurityCheck-DoubleEscaped : 4 occurrences
 	// PhanNoopSwitchCases : 1 occurrence
-	// SecurityCheck-XSS : 1 occurrence
 
 	'file_suppressions' => [
 		'includes/SpecialELNSMWAdapterUI.php' => [
 			'MediaWikiNoEmptyIfDefined' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::callAdapterService', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayProcessingPage', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayResultsPage', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displaySelectionForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::execute', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::isValidUrl'],
 			'PhanNoopSwitchCases' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayMethodForm'],
-			'PhanPluginDuplicateAdjacentStatement' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayFileUploadForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayProcessingPage', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayResults', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displaySelectionForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayUrlForm'],
-			'SecurityCheck-DoubleEscaped' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displaySelectionForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::renderDynamicField'],
-			'SecurityCheck-XSS' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayProcessingPage']
+			'PhanPluginDuplicateAdjacentStatement' => ['\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayFileUploadForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayProcessingPage', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayResults', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displaySelectionForm', '\\ELNSMWAdapterUI\\SpecialELNSMWAdapterUI::displayUrlForm']
 		],
 	],
 	// 'directory_suppressions' => ['src/directory_name' => ['PhanIssueName1', 'PhanIssueName2']] can be manually added if needed.